server, clients: use 64 bits for pids

+++ common/sl2_server_api.cpp
@@ -256,7 +256,7 @@ SL2Response sl2_conn_ping(sl2_conn *conn, uint8_t *ok)
 }
 
 SL2_EXPORT
-SL2Response sl2_conn_register_pid(sl2_conn *conn, uint32_t pid, bool tracing)
+SL2Response sl2_conn_register_pid(sl2_conn *conn, uint64_t pid, bool tracing)
 {
     DWORD txsize;
